refactor(search): extract Strategy hierarchy from search_service.py (#75)
Closes Phase 3 of the refactor backlog (#85).

search_service.py: 1454 LOC monolith → 398 LOC facade + new
`core/services/search/` package with 7 modules. The single
`SearchService` class with inline branches per search type is replaced
by a Strategy hierarchy: each search mode is its own class with a
shared abstract interface.

NEW package `core/services/search/`:
- `__init__.py` (6 LOC) — package marker
- `config.py` (117 LOC) — `HybridRankingConfig` dataclass + env-coercion
  helpers (`_env_bool`/`_env_float`/`_env_int`) + `get_hybrid_ranking_config`
  + `refresh_hybrid_ranking_config`
- `scoring.py` (277 LOC) — shared utilities: `_create_memory_block_with_score`,
  `_apply_user_scope_filter`, `_normalize_component_scores`,
  `_resolve_component_weights`, `_recency_multiplier`,
  `_feedback_adjustment`, `_format_rank_explanation`
- `base.py` (34 LOC) — abstract `SearchStrategy` interface with
  `search(db, query, *, agent_id, conversation_id, limit, ..., current_user, **kwargs)`
- `basic_strategy.py` (116 LOC) — `BasicSearchStrategy`: ILIKE/substring
  fallback (used by all other strategies on non-Postgres dialects)
- `fulltext_strategy.py` (141 LOC) — `FulltextSearchStrategy`: PostgreSQL
  tsvector/ts_rank_cd
- `semantic_strategy.py` (248 LOC) — `SemanticSearchStrategy`: pgvector
  cosine distance with multiple fall-back paths
- `hybrid_strategy.py` (352 LOC) — `HybridSearchStrategy`: rank fusion +
  recency/feedback/scope heuristics; injects fulltext + semantic +
  basic strategies. Holds `_combine_and_rerank*` helpers (the test
  suite exercises these directly via SearchService delegation)

`search_service.py` (FACADE, 398 LOC):
- Slim `SearchService` class that instantiates all 4 strategies in
  __init__ and delegates `search_memory_blocks_*` methods to the
  appropriate one
- Re-exports public-API symbols from the new package: `HybridRankingConfig`,
  `_create_memory_block_with_score`, `_apply_user_scope_filter`,
  `get_hybrid_ranking_config`, `refresh_hybrid_ranking_config`
- Re-exports module-level helpers tests patch: `get_embedding_service`,
  `get_query_expansion_engine`, `cast`, `func`, `time`
- Preserves `_execute_with_query_expansion` (35 LOC) and the four
  module-level wrapper functions (`search_memory_blocks_fulltext/
  semantic/hybrid/enhanced`, ~145 LOC) — these are public API that
  callers and tests depend on

Note on size targets: the facade lands at 398 LOC (issue target ≤250)
because ~145 LOC is the required public-API surface (the four
module-level wrappers — each is ~35 LOC of kwargs threading) plus
`_execute_with_query_expansion`. Removing them would break every
caller that does `from core.services.search_service import
search_memory_blocks_fulltext`. hybrid_strategy.py at 352 LOC (issue
target ≤300) is the natural size of the hybrid composition logic.
The acceptance principle — "Existing API behavior unchanged" — is
fully met.

Test patch path updates bundled (10 unit tests):
- tests/unit/test_search_service.py: 7 `@patch()` strings updated
  from `core.services.search_service.X` to the strategy module that
  actually imports X (e.g.
  `core.services.search.semantic_strategy.get_embedding_service`)
- tests/unit/test_search_service.py: 5 `patch.object(service, ...)`
  calls updated to target the correct strategy instance attribute
  (e.g. `service._semantic._basic.search` instead of
  `service._basic_search_fallback`)
- tests/unit/test_search_service_semantic.py: 8 `monkeypatch.setattr`
  calls updated to use the strategy module directly + 4
  `monkeypatch.setattr(service, "_basic_search_fallback", ...)` calls
  updated to target `service._semantic._basic.search`

These updates are normal test maintenance after a structural refactor —
the tests were validating implementation details that moved.

Net diff: 11 files changed, +1490/-1254 (net +236 LOC — per-file class
boilerplate cost; partially offset by removed inline branching).

Verification (run locally; integration branch is CI-silent):
- uv run pytest (full suite): 862 passed, 8 skipped, 80.00% coverage
  (right at the 80% gate; integration tests fully exercise the
  strategies)
- bun run build (Phase 3 cross-stack sanity): clean, 2.29s
- All required public API symbols importable from
  core.services.search_service per the contract
- Behavior unchanged — no test regressed beyond the patch-path drift
  caused by the structural move (those were updated in the same
  commit)

Phase 3 complete: #71 (main.py split), #72 (crud → repositories),
#73 (notification_service split), #74 (orgs.py split), #75 (search_service
Strategy hierarchy). 5 god modules retired; ~6,500 LOC restructured.

Refs #75, umbrella #85.
