+++ apps/hindsight-service/core/services/search/__init__.py
@@ -1,0 +1,6 @@
+"""
+search — strategy package for memory-block search.
+
+Public surface deliberately kept minimal; callers should import from
+``core.services.search_service`` which re-exports everything.
+"""
